Use private friend for streamable helpers
