+++ src/corecel/io/StreamableVariant.hh
@@ -7,8 +7,6 @@
 #pragma once
 
 #include <ostream>
-#include <sstream>
-#include <utility>
 #include <variant>
 
 #include "corecel/Assert.hh"
@@ -30,54 +28,27 @@ template<class T>
 struct StreamableVariant
 {
     T value;
+
+    //! Write to stream
+    friend std::ostream& operator<<(std::ostream& os,
+                                    StreamableVariant const& svar)
+    {
+        CELER_ASSUME(!svar.value.valueless_by_exception());
+        std::visit(GenericToStream{os}, svar.value);
+        return os;
+    }
+
+    //! Save as a string
+    friend std::string to_string(StreamableVariant const& svar)
+    {
+        return stream_to_string(svar);
+    }
 };
 
 //---------------------------------------------------------------------------//
 // Deduction guide
 template<class T>
 StreamableVariant(T&&) -> StreamableVariant<T>;
-
-//---------------------------------------------------------------------------//
-// IMPLEMENTATION
-//---------------------------------------------------------------------------//
-namespace detail
-{
-struct GenericToStream
-{
-    std::ostream& os;
-
-    template<class T>
-    void operator()(T&& obj) const
-    {
-        this->os << std::forward<T>(obj);
-    }
-};
-}  // namespace detail
-
-//---------------------------------------------------------------------------//
-// FREE FUNCTIONS
-//---------------------------------------------------------------------------//
-/*!
- * Write a variant object's value to a stream.
- */
-template<class T>
-inline std::ostream& operator<<(std::ostream& os,
-                                StreamableVariant<T> const& svar)
-{
-    CELER_ASSUME(!svar.value.valueless_by_exception());
-    std::visit(detail::GenericToStream{os}, svar.value);
-    return os;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Save a variant object's value to a string.
- */
-template<class T>
-inline std::string to_string(StreamableVariant<T> const& svar)
-{
-    return stream_to_string(svar);
-}
 
 //---------------------------------------------------------------------------//
 }  // namespace celeritas
